feat: add basic interpreter
Added a lex, parse, and execute module with enough power to run the
snippet of code found in the readme. Included a test script to start
automating testing of supported behavior. Wired up the test script to
the standard check command in the makefile. Added a program option for
running scripts of code from text input.