feat: add catch statement
Added a new statement for catching errors. When the catch statement is
evaluated it runs the lines of code found in its body. If an error is
thrown, then it stops running lines of code and returns the error as
the result of the statement. If no error is thrown, then it returns
null as the result of the statement.