feat: add throw statement
Added a new statement for throwing errors. When the throw statement is
evaluated it takes the result of the next expression and throws it up
the stack. If an error is thrown when evaluating that expression, then
that error is thrown. If no error is thrown, then the returned value is
thrown.