feat: add function-level scope
Added a limit to the scope of variables based on the context of the
function they were created in. Variables created at the top level of
a script have their own function scope. Scopes aren't shared with
nesting. Implementing this scope fixed a memory leak when invoking a
function variable.